Removing inner classes
