+++ onecalendar/src/main/java/com/darwindeveloper/onecalendar/clases/CalendarAdapter.java
@@ -7,7 +7,6 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 
 import com.darwindeveloper.onecalendar.R;
 
@@ -18,7 +17,7 @@ import java.util.Calendar;
  * Created by DARWIN on 1/3/2017.
  */
 
-public class CalendarAdapter extends RecyclerView.Adapter<CalendarAdapter.ViewDayHolder> {
+public class CalendarAdapter extends RecyclerView.Adapter<ViewDayHolder> {
 
     private Context context;
     private ArrayList<Day> dias;
@@ -81,37 +80,6 @@ public class CalendarAdapter extends RecyclerView.Adapter<CalendarAdapter.ViewDa
     @Override
     public int getItemCount() {
         return dias.size();
-    }
-
-
-    public class ViewDayHolder extends RecyclerView.ViewHolder {
-
-        Button dia;
-
-        public ViewDayHolder(View itemView) {
-            super(itemView);
-            dia = (Button) itemView.findViewById(R.id.textViewDay);
-        }
-    }
-
-
-    public interface DayOnClickListener {
-        /**
-         * un objeto de tipo day para obtener la fecha (año,mes,dia) con un objeto calendar
-         * <p>
-         * otros metodos como setBackgroundColor(int backgroundColor) y getBackgroundColor() color del fondo del numero de dia del mes
-         * setTextColor(int textColor) y getTextColor() color del texto numero de dia del mes
-         *
-         * @param day
-         */
-        void dayOnClick(Day day, int position);
-
-        /**
-         * similar a dayOnClick solo que este se ejecuta cuando haya un clic prolongado
-         *
-         * @param day
-         */
-        void dayOnLongClik(Day day, int position);
     }
