+++ onecalendar/src/main/java/com/darwindeveloper/onecalendar/clases/DayOnClickListener.java
@@ -1,0 +1,20 @@
+package com.darwindeveloper.onecalendar.clases;
+
+public interface DayOnClickListener {
+    /**
+     * un objeto de tipo day para obtener la fecha (año,mes,dia) con un objeto calendar
+     * <p>
+     * otros metodos como setBackgroundColor(int backgroundColor) y getBackgroundColor() color del fondo del numero de dia del mes
+     * setTextColor(int textColor) y getTextColor() color del texto numero de dia del mes
+     *
+     * @param day
+     */
+    void dayOnClick(Day day, int position);
+
+    /**
+     * similar a dayOnClick solo que este se ejecuta cuando haya un clic prolongado
+     *
+     * @param day
+     */
+    void dayOnLongClik(Day day, int position);
+}
